US 03.06.01 As an administrator, I want to be able to browse images. US 03.03.01 As an administrator, I want to be able to remove images.
Finished implementing US 03.06.01 and  US 03.03.01
also fixed most recent bug where the profile picture button was not working.

NOTE: These two user stories have some problems because they do not track where the images come from. They only give the ability to delete and view images. Meaning that if you delete an image the User that had that image as a path will not be updated.